updated user routes to secure based on active session

+++ controllers/user_controller.js
@@ -1,4 +1,5 @@
 const express = require('express')
+const bcrypt = require('bcryptjs');
 const router = express.Router()
 
 const db = require('../models')
@@ -22,7 +23,17 @@ router.get('/', async(req, res, next) => {
 router.get('/:id/edit', async(req, res, next) => {
     try {
         const user = await db.User.findById(req.params.id)
-        res.render("./users/edit.ejs", {user: user});
+        if(req.session.currentUser) {
+            if(req.session.currentUser.id === req.params.id) {
+                res.render("./users/edit.ejs", {user: user});
+            }
+            else {
+                res.redirect(`/users/${req.params.id}`);
+            }
+        }
+        else {
+            res.redirect('/');
+        }
     }
     catch(err) {
         console.log("Error in user edit: " + err);
@@ -34,7 +45,14 @@ router.get('/:id/edit', async(req, res, next) => {
 router.get('/:id', async(req, res, next) => {
     try {
         const user = await db.User.findById(req.params.id).populate('parks').populate('badges');
-        let editOK = true; // This will be set if show :id === session user id
+        let editOK = false;
+
+        // a user should only be able to edit their own profile
+        if(req.session.currentUser) {
+            if(req.session.currentUser.id === req.params.id) {
+                editOK = true; 
+            }
+        }
 
         res.render("./users/show.ejs", {user: user, editOK: editOK});
     }
@@ -47,15 +65,26 @@ router.get('/:id', async(req, res, next) => {
 // put route (updates user info)
 router.put('/:id', async (req, res, next) => {
     try {
-        if(req.body.password === '') {
-            let updatedUser = await db.User.findByIdAndUpdate(req.params.id, {name: req.body.name, avatar: req.body.avatar, email: req.body.email});
+        if(req.session.currentUser) {
+            if(req.session.currentUser.id === req.params.id) {
+                if(req.body.password === '') {
+                    let updatedUser = await db.User.findByIdAndUpdate(req.params.id, {name: req.body.name, avatar: req.body.avatar, email: req.body.email});
+                }
+                else {
+                    const salt = await bcrypt.genSalt(12);
+                    const hash = await bcrypt.hash(req.body.password, salt);
+                    req.body.password = hash;
+                    let updatedUser = await db.User.findByIdAndUpdate(req.params.id, {name: req.body.name, avatar: req.body.avatar, email: req.body.email, password: req.body.password});
+                }
+                res.redirect(`/users/${req.params.id}`);
+            }
+            else {
+                res.redirect(`/users/${req.params.id}`);
+            }
         }
         else {
-            //TODO: Add bcrypt stuff before updating password field
-            let updatedUser = await db.User.findByIdAndUpdate(req.params.id, {name: req.body.name, avatar: res.body.avatar, email: req.body.email, password: req.body.password});
+            res.redirect('/');
         }
-        res.redirect(`/users/${req.params.id}`);
-        
     }
     catch(err) {
         console.log("Error in user put: " + err);
@@ -66,8 +95,18 @@ router.put('/:id', async (req, res, next) => {
 // delete route
 router.delete('/:id', async(req, res, next) => {
     try {
-        let deletedUser = await db.User.findByIdAndDelete(req.params.id);
-        res.redirect("/");
+        if(req.session.currentUser) {
+            if(req.session.currentUser.id === req.params.id) {
+                let deletedUser = await db.User.findByIdAndDelete(req.params.id);
+                res.redirect("/");
+            }
+            else {
+                res.redirect(`/users/${req.params.id}`);
+            }
+        }
+        else {
+            res.redirect('/');
+        }
     }
     catch(err) {
         console.log("Error in delete put: " + err);
